3rd Commit - Fixed operation logic for division and multiplication

+++ CalculatorPrototype/Models/Calculator.cs
@@ -9,7 +9,7 @@ namespace CalculatorPrototype.Models
     {
         public int FirstInput { get; set; }
         public int SecondInput { get; set; }
-        public int Result { get; set; }
+        public decimal Result { get; set; }
         public int Operator { get; set; }
     }
 }
